--- src/main/java/org/c4marathon/assignment/domain/user/dto/JoinResponseDto.java
@@ -1,5 +1,0 @@
-package org.c4marathon.assignment.domain.user.dto;
-
-
-public record JoinResponseDto(String responseMsg) {
-}
