--- src/main/java/org/c4marathon/assignment/domain/user/dto/LoginResponseDto.java
@@ -1,6 +1,0 @@
-package org.c4marathon.assignment.domain.user.dto;
-
-
-public record LoginResponseDto(String responseMsg) {
-}
-
