+++ src/main/java/org/c4marathon/assignment/domain/user/dto/response/JoinResponseDto.java
@@ -1,0 +1,5 @@
+package org.c4marathon.assignment.domain.user.dto.response;
+
+
+public record JoinResponseDto(String responseMsg) {
+}
