+++ src/main/java/org/c4marathon/assignment/domain/user/dto/response/LoginResponseDto.java
@@ -1,0 +1,6 @@
+package org.c4marathon.assignment.domain.user.dto.response;
+
+
+public record LoginResponseDto(String responseMsg) {
+}
+
