+++ src/main/java/org/c4marathon/assignment/domain/user/dto/response/UserErrDto.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain.user.dto;
+package org.c4marathon.assignment.domain.user.dto.response;
 
 import org.springframework.http.HttpStatus;
